Dev be

See merge request s12-final/S12P31B105!114

+++ backend/src/main/java/com/ssafy/hellojob/domain/companyanalysis/entity/CompanyAnalysis.java
@@ -20,7 +20,7 @@ public class CompanyAnalysis {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "company_analysis_id", nullable = false)
-    private Long companyAnalysisId;
+    private Integer companyAnalysisId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
