+++ backend/src/main/java/com/ssafy/hellojob/domain/companyanalysis/repository/DartAnalysisRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DartAnalysisRepository extends JpaRepository<DartAnalysis, Long> {
+public interface DartAnalysisRepository extends JpaRepository<DartAnalysis, Integer> {
 
 }
