+++ backend/src/main/java/com/ssafy/hellojob/domain/coverletter/service/CoverLetterService.java
@@ -49,7 +49,7 @@ public class CoverLetterService {
         if (requestDto.getJobRoleAnalysisId() == null) {
             jobRoleSnapshot = null;
         } else {
-            JobRoleAnalysis jobRoleAnalysis = jobRoleAnalysisRepository.findById(requestDto.getJobRoleAnalysisId().longValue())
+            JobRoleAnalysis jobRoleAnalysis = jobRoleAnalysisRepository.findById(requestDto.getJobRoleAnalysisId())
                     .orElseThrow(() -> new BaseException(ErrorCode.JOB_ROLE_ANALYSIS_NOT_FOUND));
 
             jobRoleSnapshot = jobRoleSnapshotService.copyJobRoleAnalysis(companyName, jobRoleAnalysis);
